+++ src/App.js
@@ -1,9 +1,9 @@
 import React from 'react';
-import Home from './pages/Home';
-import NewProjet from './pages/NewProjet';
-import VueEnsemble from './pages/VueEnsemble';
-import Projet from './pages/Projet';
-import Profil from './pages/Profil';
+import Home from './pages/Home/Home';
+import NewProjet from './pages/NewProjet/NewProjet';
+import VueEnsemble from './pages/VueEnsemble/VueEnsemble';
+import Projet from './pages/Projet/Projet';
+import Profil from './pages/Profil/Profil';
 import { BrowserRouter, Switch, Route } from 'react-router-dom';
