+++ src/pages/Home/Home.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { Button } from 'reactstrap';
 import 'bootstrap/dist/css/bootstrap.css';
-import '../Home.css';
+import './Home.css';
 import { Link } from 'react-router-dom';
 import NavBar from '../NavBar';
 
@@ -21,6 +21,7 @@ class Home extends Component {
         <Link to="/VueEnsemble">
           <Button color="primary" size="lg" className="ConsultationProjet">Consultation Projets</Button>
         </Link>
+        <NavBar/>
       </div>
     );
   }
